Got V2 working properly now. Can extend functionality, Renames the fp 8 folder for consistency sake
