+++ src/csurf_faderport_8/csurf_fp_8_channel_manager.hpp
