+++ src/csurf_faderport_8/csurf_fp_8_main.cpp
@@ -1,6 +1,6 @@
 /*
 ** reaper_csurf
-** FaderPort V2 support
+** FaderPort support
 ** Copyright (C) 2007-2008 Cockos Incorporated
 ** License: LGPL.
 */
@@ -12,34 +12,36 @@
 #include <WDL/ptrlist.h>
 #include <reaper_plugin_functions.h>
 #include "../controls/csurf_button.hpp"
-#include "../shared/csurf.h"
-#include "../shared/csurf_transport_manager.hpp"
-#include "../shared/csurf_utils.hpp"
-#include "../csurf_faderport/csurf_fp_8_last_touched_fx_manager.hpp"
-#include "csurf_fp_v2_session_manager.cpp"
-#include "csurf_fp_v2_track_manager.hpp"
-#include "csurf_fp_v2_automation_manager.cpp"
-#include "csurf_fp_v2_general_control_manager.cpp"
-#include "csurf_fp_v2_navigator.hpp"
-#include "csurf_fp_v2_ui_init.hpp"
 #include "../resource.h"
+#include "../shared/csurf.h"
+#include "../shared/csurf_transport_manager.cpp"
+#include "../shared/csurf_utils.hpp"
+#include "csurf_fp_8_session_manager.cpp"
+#include "csurf_fp_8_mix_manager.cpp"
+#include "csurf_fp_8_automation_manager.cpp"
+#include "csurf_fp_8_general_control_manager.cpp"
+#include "csurf_fp_8_last_touched_fx_manager.hpp"
+#include "csurf_fp_8_fader_manager.hpp"
+#include "csurf_fp_8_ui_init.hpp"
+#include "csurf_fp_8_navigator.hpp"
 
 extern HWND g_hwnd;
 extern REAPER_PLUGIN_HINSTANCE g_hInst;
 
-class CSurf_FaderPortV2 : public IReaperControlSurface
+class CSurf_FaderPort : public IReaperControlSurface
 {
   int m_midi_in_dev, m_midi_out_dev;
   midi_Output *m_midiout;
   midi_Input *m_midiin;
 
   CSurf_Context *context;
-  CSurf_FP_V2_SessionManager *sessionManager;
-  CSurf_FP_V2_Navigator *trackNavigator;
-  CSurf_FP_V2_TrackManager *trackManager;
+  CSurf_FP_8_SessionManager *sessionManager;
+  CSurf_FP_8_FaderManager *faderManager;
+  CSurf_FP_8_Navigator *trackNavigator;
+  CSurf_FP_8_MixManager *mixManager;
   CSurf_TransportManager *transportManager;
-  CSurf_FP_V2_AutomationManager *automationManager;
-  CSurf_FP_V2_GeneralControlManager *generalControlManager;
+  CSurf_FP_8_AutomationManager *automationManager;
+  CSurf_FP_8_GeneralControlManager *generalControlManager;
   CSurf_FP_8_LastTouchedFXManager *lastTouchedFxManager;
 
   std::vector<CSurf_FP_8_Track *> tracks;
@@ -57,9 +59,18 @@ class CSurf_FaderPortV2 : public IReaperControlSurface
     /**
      * Fader values
      */
-    if (evt->midi_message[0] == FADER_1)
+    if (evt->midi_message[0] >= FADER_1 && evt->midi_message[0] <= FADER_16)
     {
-      trackManager->HandleFaderMove(evt->midi_message[2], evt->midi_message[1]);
+      if (
+          (context->GetNbChannels() == 7 && context->GetLastTouchedFxMode() && evt->midi_message[0] == FADER_8) ||
+          (context->GetNbChannels() == 15 && context->GetLastTouchedFxMode() && evt->midi_message[0] == FADER_16))
+      {
+        lastTouchedFxManager->HandleFaderMove(evt->midi_message[2], evt->midi_message[1]);
+      }
+      else
+      {
+        faderManager->HandleFaderMove(evt->midi_message[0] - FADER_1, evt->midi_message[2], evt->midi_message[1]);
+      }
     }
 
     /**
@@ -69,7 +80,7 @@ class CSurf_FaderPortV2 : public IReaperControlSurface
     {
       if (evt->midi_message[1] == ENCODER_PAN)
       {
-        // generalControlManager->HandleEncoderChange(evt->midi_message[2]);
+        generalControlManager->HandleEncoderChange(evt->midi_message[2]);
       }
       if (evt->midi_message[1] == ENCODER_NAV)
       {
@@ -85,24 +96,75 @@ class CSurf_FaderPortV2 : public IReaperControlSurface
       /**
        * Fader Touch
        */
-      if (evt->midi_message[1] == FADER_TOUCH_1)
+      if (evt->midi_message[1] >= FADER_TOUCH_1 && evt->midi_message[1] <= FADER_TOUCH_16)
       {
-        trackManager->HandleFaderTouch(evt->midi_message[2]);
+        faderManager->HandleFaderTouch(evt->midi_message[1] - FADER_TOUCH_1, evt->midi_message[2]);
       }
 
       /**
+       * Track Select Buttons
+       */
+      else if (evt->midi_message[1] >= BTN_SELECT_1 && evt->midi_message[1] <= BTN_SELECT_8)
+      {
+        faderManager->HandleSelectClick(evt->midi_message[1] - BTN_SELECT_1, evt->midi_message[2]);
+      }
+      else if (evt->midi_message[1] == BTN_SELECT_9)
+      {
+        faderManager->HandleSelectClick(8, evt->midi_message[2]);
+      }
+      else if (evt->midi_message[1] >= BTN_SELECT_10 && evt->midi_message[1] <= BTN_SELECT_16)
+      {
+        faderManager->HandleSelectClick(evt->midi_message[1] - 0x20 + 8, evt->midi_message[2]);
+      }
+      /**
        * Track Mute Buttons
        */
-      else if (evt->midi_message[1] == BTN_MUTE_1)
+      else if (evt->midi_message[1] >= BTN_MUTE_1 && evt->midi_message[1] <= BTN_MUTE_8)
       {
-        trackManager->HandleMuteClick(1, evt->midi_message[2]);
+        faderManager->HandleMuteClick(evt->midi_message[1] - BTN_MUTE_1, evt->midi_message[2]);
+      }
+      else if (evt->midi_message[1] >= BTN_MUTE_9 && evt->midi_message[1] <= BTN_MUTE_16)
+      {
+        faderManager->HandleMuteClick(evt->midi_message[1] - BTN_MUTE_9 + 8, evt->midi_message[2]);
       }
       /**
        * Track Solo Buttons
        */
-      else if (evt->midi_message[1] == BTN_SOLO_1)
+      else if (evt->midi_message[1] >= BTN_SOLO_1 && evt->midi_message[1] <= BTN_SOLO_8)
       {
-        trackManager->HandleSoloClick(evt->midi_message[1] - BTN_SOLO_1, evt->midi_message[2]);
+        faderManager->HandleSoloClick(evt->midi_message[1] - BTN_SOLO_1, evt->midi_message[2]);
+      }
+      else if (evt->midi_message[1] >= BTN_SOLO_9 && evt->midi_message[1] <= BTN_SOLO_14 && evt->midi_message[1] != BTN_SOLO_12 && evt->midi_message[1] != ENCODER_CLICK_NAV)
+      {
+        faderManager->HandleSoloClick(evt->midi_message[1] - BTN_SOLO_9 + 8, evt->midi_message[2]);
+      }
+      else if (evt->midi_message[1] == BTN_SOLO_14 || evt->midi_message[1] == BTN_SOLO_15)
+      {
+        faderManager->HandleSoloClick(evt->midi_message[1] == BTN_SOLO_14 ? 14 : 15, evt->midi_message[2]);
+      }
+      else if (evt->midi_message[1] == BTN_SOLO_12)
+      {
+        faderManager->HandleSoloClick(12, evt->midi_message[2]);
+      }
+
+      /**
+       * Fader Buttons
+       */
+      else if (evt->midi_message[1] == BTN_TRACK)
+      {
+        faderManager->HandleTrackButtonClick(evt->midi_message[2]);
+      }
+      else if (evt->midi_message[1] == BTN_EDIT_PLUGINS)
+      {
+        faderManager->HandlePluginsButtonClick(evt->midi_message[2]);
+      }
+      else if (evt->midi_message[1] == BTN_SEND)
+      {
+        faderManager->HandleSendButtonClick(evt->midi_message[2]);
+      }
+      else if (evt->midi_message[1] == BTN_PAN)
+      {
+        faderManager->HandlePanButtonClick(evt->midi_message[2]);
       }
 
       /**
@@ -134,20 +196,60 @@ class CSurf_FaderPortV2 : public IReaperControlSurface
       }
 
       /**
+       * Mix Management
+       */
+      else if (evt->midi_message[1] == BTN_AUDIO)
+      {
+        mixManager->HandleMixAudioButton(evt->midi_message[2]);
+      }
+      else if (evt->midi_message[1] == BTN_VI)
+      {
+        mixManager->HandleMixViButton(evt->midi_message[2]);
+      }
+      else if (evt->midi_message[1] == BTN_BUS)
+      {
+        mixManager->HandleMixBusButton(evt->midi_message[2]);
+      }
+      else if (evt->midi_message[1] == BTN_VCA)
+      {
+        mixManager->HandleMixVcaButton(evt->midi_message[2]);
+      }
+      else if (evt->midi_message[1] == BTN_ALL)
+      {
+        mixManager->HandleMixAllButton(evt->midi_message[2]);
+      }
+      else if (evt->midi_message[1] == BTN_SHIFT_RIGHT)
+      {
+        mixManager->HandleShiftButton(evt->midi_message[2]);
+      }
+
+      /**
        * General Control Management
        */
       else if (evt->midi_message[1] == ENCODER_CLICK_PAN)
       {
-        // generalControlManager->HandleEncoderClick(evt->midi_message[2]);
+        generalControlManager->HandleEncoderClick(evt->midi_message[2]);
       }
       else if (evt->midi_message[1] == BTN_ARM)
       {
-        // generalControlManager->HandleArmButton(evt->midi_message[2]);
+        generalControlManager->HandleArmButton(evt->midi_message[2]);
         sessionManager->Refresh();
+      }
+      else if (evt->midi_message[1] == BTN_SOLO_CLEAR)
+      {
+        generalControlManager->HandleSoloClearButton(evt->midi_message[2]);
+      }
+      else if (evt->midi_message[1] == BTN_MUTE_CLEAR)
+      {
+        generalControlManager->HandleMuteClearButton(evt->midi_message[2]);
       }
       else if (evt->midi_message[1] == BTN_BYPASS)
       {
-        // generalControlManager->HandleBypassButton(evt->midi_message[2]);
+        generalControlManager->HandleBypassButton(evt->midi_message[2]);
+      }
+      else if (evt->midi_message[1] == BTN_MACRO)
+      {
+        generalControlManager->HandleMacroButton(evt->midi_message[2]);
       }
       else if (evt->midi_message[1] == BTN_LINK)
       {
@@ -162,9 +264,25 @@ class CSurf_FaderPortV2 : public IReaperControlSurface
       /**
        * Automation Management
        */
+      else if (evt->midi_message[1] == BTN_LATCH)
+      {
+        automationManager->HandleLatchButton(evt->midi_message[2]);
+      }
+      else if (evt->midi_message[1] == BTN_TRIM)
+      {
+        automationManager->HandleTrimButton(evt->midi_message[2]);
+      }
+      else if (evt->midi_message[1] == BTN_OFF)
+      {
+        automationManager->HandleOffButton(evt->midi_message[2]);
+      }
       else if (evt->midi_message[1] == BTN_TOUCH)
       {
         automationManager->HandleTouchButton(evt->midi_message[2]);
+      }
+      else if (evt->midi_message[1] == BTN_WRITE)
+      {
+        automationManager->HandleWriteButton(evt->midi_message[2]);
       }
       else if (evt->midi_message[1] == BTN_WRITE)
       {
@@ -184,11 +302,15 @@ class CSurf_FaderPortV2 : public IReaperControlSurface
       }
       else if (evt->midi_message[1] == BTN_ZOOM)
       {
-        // sessionManager->HandleZoomButton(evt->midi_message[2]);
+        sessionManager->HandleZoomButton(evt->midi_message[2]);
       }
       else if (evt->midi_message[1] == BTN_SCROLL)
       {
         sessionManager->HandleScrollButton(evt->midi_message[2]);
+      }
+      else if (evt->midi_message[1] == BTN_BANK)
+      {
+        sessionManager->HandleBankButton(evt->midi_message[2]);
       }
       else if (evt->midi_message[1] == BTN_MASTER)
       {
@@ -222,7 +344,7 @@ class CSurf_FaderPortV2 : public IReaperControlSurface
   }
 
 public:
-  CSurf_FaderPortV2(int indev, int outdev, int *errStats)
+  CSurf_FaderPort(int indev, int outdev, int *errStats)
   {
     (void)indev;
     (void)outdev;
@@ -232,7 +354,7 @@ public:
      *
      */
     mINI::INIStructure ini;
-    readAndCreateIni(ini);
+    readAndCreateIni(ini, FP_8);
 
     errStats = 0;
     m_midi_in_dev = stoi(ini["surface"]["midiin"]);
@@ -256,12 +378,13 @@ public:
     }
     lastTouchedFxTrack = new CSurf_FP_8_Track(context->GetNbChannels() - 1, context, m_midiout);
 
-    trackNavigator = new CSurf_FP_V2_Navigator(context);
-    sessionManager = new CSurf_FP_V2_SessionManager(context, trackNavigator, m_midiout);
-    trackManager = new CSurf_FP_V2_TrackManager(context, trackNavigator, m_midiout);
+    trackNavigator = new CSurf_FP_8_Navigator(context);
+    sessionManager = new CSurf_FP_8_SessionManager(context, trackNavigator, m_midiout);
+    faderManager = new CSurf_FP_8_FaderManager(context, trackNavigator, m_midiout);
+    mixManager = new CSurf_FP_8_MixManager(context, trackNavigator, faderManager, m_midiout);
     transportManager = new CSurf_TransportManager(context, m_midiout);
-    automationManager = new CSurf_FP_V2_AutomationManager(context, m_midiout);
-    generalControlManager = new CSurf_FP_V2_GeneralControlManager(context, trackNavigator, m_midiout);
+    automationManager = new CSurf_FP_8_AutomationManager(context, m_midiout);
+    generalControlManager = new CSurf_FP_8_GeneralControlManager(context, trackNavigator, faderManager, m_midiout);
     lastTouchedFxManager = new CSurf_FP_8_LastTouchedFXManager(lastTouchedFxTrack, context, m_midiout);
 
     if (errStats)
@@ -290,7 +413,7 @@ public:
     }
   }
 
-  ~CSurf_FaderPortV2()
+  ~CSurf_FaderPort()
   {
     if (m_midiout)
     {
@@ -311,12 +434,12 @@ public:
 
   const char *GetTypeString()
   {
-    return "REASONUS_FADERPORT";
+    return "REASONUS_FADERPORT_8";
   }
 
   const char *GetDescString()
   {
-    snprintf(configtmp, 100, "ReaSonus FaderPort (dev %d, %d)", m_midi_in_dev, m_midi_out_dev);
+    snprintf(configtmp, 100, "ReaSonus FaderPort 8 & 16 (dev %d, %d)", m_midi_in_dev, m_midi_out_dev);
     return configtmp;
   }
 
@@ -355,7 +478,7 @@ public:
       DWORD now = GetTickCount();
       if ((now - surface_update_lastrun) >= 100)
       {
-        trackManager->UpdateTracks();
+        faderManager->UpdateTracks();
         if (context->GetLastTouchedFxMode())
         {
           lastTouchedFxManager->UpdateTrack();
@@ -392,18 +515,18 @@ public:
   }
 };
 
-static IReaperControlSurface *createFuncV2(const char *type_string, const char *configString, int *errStats)
+static IReaperControlSurface *createFunc(const char *type_string, const char *configString, int *errStats)
 {
   (void)type_string;
   int parms[4];
   parseParms(configString, parms);
 
-  return new CSurf_FaderPortV2(parms[2], parms[3], errStats);
+  return new CSurf_FaderPort(parms[2], parms[3], errStats);
 }
 
-reaper_csurf_reg_t csurf_faderport_v2_reg = {
-    "REASONUS_FADERPORT_V2",
-    "ReaSonus FaderPort V2",
-    createFuncV2,
-    CSURF_FP_V2_UI_INIT::CreateInitDialog,
+reaper_csurf_reg_t csurf_faderport_8_reg = {
+    "REASONUS_FADERPORT_8",
+    "ReaSonus FaderPort 8 & 16",
+    createFunc,
+    CSURF_FP_8_UI_INIT::CreateInitDialog,
 };
