+++ src/csurf_faderport_8/csurf_fp_8_plugin_control_manager.cpp
