+++ src/csurf_faderport_8/csurf_fp_8_ui_functions.cpp
@@ -40,7 +40,7 @@ namespace CSURF_FP_UI_FUNCTIONS
         g_querying_action = false;
         if (actionId > 0)
         {
-            static mINI::INIFile file(GetReaSonusIniPath());
+            static mINI::INIFile file(GetReaSonusIniPath(FP_8));
             ini["Functions"][functionsDlgSelectedFunction] = std::to_string(actionId);
             file.write(ini);
             PopulateActionFields(s_hwndDlg);
@@ -176,7 +176,7 @@ namespace CSURF_FP_UI_FUNCTIONS
 
     void ShowFunctionsDialog()
     {
-        mINI::INIFile file(GetReaSonusIniPath());
+        mINI::INIFile file(GetReaSonusIniPath(FP_8));
         file.read(ini);
 
         if (s_hwndReaSonusFunctionsDlg == NULL)
