+++ src/csurf_faderport_v2/CMakeLists.txt
@@ -1,6 +1,6 @@
 target_sources(${PROJECT_NAME}
     PRIVATE
-        csurf_faderport_v2.cpp
+        csurf_fp_v2_main.cpp
         csurf_fp_v2_automation_manager.cpp
         csurf_fp_v2_general_control_manager.cpp
         csurf_fp_v2_navigator.cpp
