+++ src/csurf_faderport_v2/csurf_fp_v2_session_manager.cpp
@@ -56,7 +56,7 @@ protected:
 
     void handleFunctionKey(std::string key)
     {
-        mINI::INIFile file(GetReaSonusIniPath());
+        mINI::INIFile file(GetReaSonusIniPath(FP_V2));
         mINI::INIStructure ini;
         file.read(ini);
