+++ src/main.cpp
@@ -7,8 +7,9 @@
 #include "actions/close_all_floating_fx_windows.hpp"
 #include "resource.h"
 
-extern reaper_csurf_reg_t csurf_faderport_reg;
-extern reaper_csurf_reg_t csurf_faderport_v2_reg;
+extern reaper_csurf_reg_t
+    csurf_faderport_8_reg,
+    csurf_faderport_v2_reg;
 
 REAPER_PLUGIN_HINSTANCE g_hInst; // used for dialogs, if any
 HWND g_hwnd;
@@ -55,7 +56,7 @@ extern "C"
     SHOW_REASONUS_FUNCTION_WINDOW::Register();
     SHOW_REASONUS_FILTERS_WINDOW::Register();
     CLOSE_ALL_FLOATING_FX_WINDOWS::Register();
-    reaper_plugin_info->Register("csurf", &csurf_faderport_reg);
+    reaper_plugin_info->Register("csurf", &csurf_faderport_8_reg);
     reaper_plugin_info->Register("csurf", &csurf_faderport_v2_reg);
     return 1;
   }
