+++ src/res.rc
@@ -182,15 +182,12 @@ STYLE DS_CONTROL | WS_CHILD
 CAPTION "ReaSonus FaderPort Native"
 FONT 8, "MS Shell Dlg"
 BEGIN
-    GROUPBOX        "Surface",IDC_GROUP_SURFACE,10,5,190,43
-    COMBOBOX        IDC_COMBO_SURFACE,55,25,135,30,CBS_DROPDOWNLIST | WS_VSCROLL | WS_TABSTOP
-    LTEXT           "Surface",IDC_STATIC,20,26,28,10
-    GROUPBOX        "MIDI Connections",IDC_GROUP_MIDI_DEVICES,10,65,190,63
-    COMBOBOX        IDC_COMBO_MIDI_IN,55,85,135,30,CBS_DROPDOWNLIST | WS_VSCROLL | WS_TABSTOP
-    LTEXT           "MIDI In",IDC_STATIC,20,86,28,10
-    COMBOBOX        IDC_COMBO_MIDI_OUT,55,105,135,30,CBS_DROPDOWNLIST | WS_VSCROLL | WS_TABSTOP
-    LTEXT           "MIDI Out",IDC_STATIC,20,106,31,10
-    GROUPBOX        "Settings",IDC_GROUP_SETTINGS,210,5,140,128
+    GROUPBOX        "MIDI Connections",IDC_GROUP_MIDI_DEVICES,10,5,190,63
+    COMBOBOX        IDC_COMBO_MIDI_IN,55,25,135,30,CBS_DROPDOWNLIST | WS_VSCROLL | WS_TABSTOP
+    LTEXT           "MIDI In",IDC_STATIC,20,26,28,10
+    COMBOBOX        IDC_COMBO_MIDI_OUT,55,45,135,30,CBS_DROPDOWNLIST | WS_VSCROLL | WS_TABSTOP
+    LTEXT           "MIDI Out",IDC_STATIC,20,46,31,10
+    GROUPBOX        "Settings",IDC_GROUP_SETTINGS,210,5,140,63
     CONTROL         " Mute/Solo Momentary",IDC_CHECK_INIT_MUTE_MOMENTARY,"Button",BS_AUTOCHECKBOX | WS_TABSTOP,220,25,110,10
     PUSHBUTTON      "Documentation",IDC_BUTTON_DOCUMENTTION,10,185,60,20
