+++ src/shared/csurf_daw.hpp
@@ -23,6 +23,8 @@ public:
     static std::string GetTrackInputName(MediaTrack *media_track);
     static std::string GetTrackMonitorMode(MediaTrack *media_track);
     static std::string GetTrackRecordingMode(MediaTrack *media_track);
+    static bool GetTrackFxBypassed(MediaTrack *media_track);
+    static void ToggleTrackFxBypass(MediaTrack *media_track);
 
     static double GetTrackPeakInfo(MediaTrack *media_track);
     static int GetTrackSurfacePeakInfo(MediaTrack *media_track);
